fix frontend api fetching for patient registration
