+++ frontend/api/general.ts
@@ -1,4 +1,4 @@
-import { get } from "./fetch"
+import { get, post } from "./fetch"
 
 const API_URL = process.env.EXPO_PUBLIC_API_URL ?? "http://localhost:8080" // for testing, when env not loaded
 
@@ -8,3 +8,27 @@ export const getHealth = async () => {
   })
   return data
 }
+
+export const getPatientData = async (patientId: string) => {
+  const data = await get({
+    url: `${API_URL}/api/patients/${patientId}`,
+  })
+  return data
+}
+
+export const registerPatient = async (patientData: {
+  email: string;
+  password: string;
+  firstName: string;
+  lastName: string;
+  phoneNumber: string;
+  deviceToken: string[];
+}) => {
+  const data = await post({
+    url: `${API_URL}/api/patients/register`,
+    body: patientData,
+  })
+  return data
+};
+
+
